add:createName method to controller
