--- src/main/java/com/example/mybatis0603/NameController.java
@@ -1,20 +1,0 @@
-package com.example.mybatis0603;
-
-import org.springframework.web.bind.annotation.GetMapping;
-import org.springframework.web.bind.annotation.RestController;
-
-import java.util.List;
-
-@RestController
-public class NameController {
-    public final NameMapper nameMapper;
-
-    public NameController(NameMapper nameMapper) {
-        this.nameMapper = nameMapper;
-    }
-
-    @GetMapping("/name")
-    public List<Name> names() {
-        return nameMapper.findAll();
-    }
-}
