+++ src/main/java/com/example/mybatis0603/NameMapper.java
@@ -1,6 +1,9 @@
 package com.example.mybatis0603;
 
+import com.example.mybatis0603.controller.Name;
+import org.apache.ibatis.annotations.Insert;
 import org.apache.ibatis.annotations.Mapper;
+import org.apache.ibatis.annotations.Param;
 import org.apache.ibatis.annotations.Select;
 
 import java.util.List;
@@ -13,4 +16,7 @@ public interface NameMapper {
 
     @Select("SELECT * FROM anime WHERE id = #{id}")
     Optional<Name> findById(int id);
+
+    @Insert("INSERT INTO anime (characterName) VAKUE (#{name}) ")
+    int insert(@Param("name") String name);
 }
